requirements.txt
fastapi
uvicorn
sqlalchemy
playwright
groq
python-dotenv
jinja2
